Start implementing runtime APIs in Millau runtime
